+++ pimcore/lib/Pimcore/Bundle/AdminBundle/Controller/Admin/UserController.php
@@ -796,6 +796,8 @@ class UserController extends AdminController implements EventedControllerInterfa
      */
     public function renew2FaSecretAction(Request $request)
     {
+        $this->checkCsrfToken($request);
+
         $user = $this->getAdminUser();
         $proxyUser = $this->getAdminUser(true);
