+++ pimcore/lib/Pimcore/Bundle/AdminBundle/Controller/AdminController.php
@@ -19,6 +19,7 @@ use Pimcore\Bundle\AdminBundle\Security\User\TokenStorageUserResolver;
 use Pimcore\Bundle\AdminBundle\Security\User\User as UserProxy;
 use Pimcore\Controller\Controller;
 use Pimcore\Model\User;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Event\FilterControllerEvent;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\Serializer\Encoder\DecoderInterface;
@@ -193,5 +194,13 @@ abstract class AdminController extends Controller implements AdminControllerInte
         $translator = $this->get('translator');
 
         return $translator->trans($id, $parameters, $domain, $locale);
+    }
+
+    /**
+     * @param Request $request
+     */
+    public function checkCsrfToken(Request $request) {
+        $csrfCheck = $this->container->get('Pimcore\Bundle\AdminBundle\EventListener\CsrfProtectionListener');
+        $csrfCheck->checkCsrfToken($request);
     }
 }
